Refactor CLI suggest to new module
